pyproject.toml
[project]
name = "django-ms-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "django>=5.2.2",
    "django-otp>=1.6.0",
    "django-two-factor-auth>=1.17.0",
    "phonenumbers>=9.0.7",
    "qrcode>=7.4.2",
]
